refactor tests, reorganize test packages
