+++ app/src/androidTest/java/com/canli/oya/traininventoryroom/fragments/BrandListFragmentTest.kt
@@ -1,4 +1,4 @@
-package com.canli.oya.traininventoryroom.ui.categories
+package com.canli.oya.traininventoryroom.fragments
 
 import android.content.Context
 import android.os.Bundle
@@ -15,49 +15,38 @@ import androidx.test.espresso.matcher.ViewMatchers.*
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.filters.MediumTest
 import com.canli.oya.traininventoryroom.R
-import com.canli.oya.traininventoryroom.data.CategoryEntry
-import com.canli.oya.traininventoryroom.data.source.FakeCategoryDataSource
+import com.canli.oya.traininventoryroom.data.BrandEntry
+import com.canli.oya.traininventoryroom.data.source.FakeBrandDataSource
 import com.canli.oya.traininventoryroom.data.source.IBrandCategoryDataSource
+import com.canli.oya.traininventoryroom.data.source.sampleBrandList
 import com.canli.oya.traininventoryroom.di.AndroidTestApplication
 import com.canli.oya.traininventoryroom.di.TestComponent
-import com.canli.oya.traininventoryroom.ui.main.Navigator
+import com.canli.oya.traininventoryroom.ui.brands.BrandListFragment
 import com.canli.oya.traininventoryroom.utils.clickOnChildWithId
 import com.canli.oya.traininventoryroom.utils.isGone
 import com.canli.oya.traininventoryroom.utils.isVisible
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runBlockingTest
-import org.junit.After
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
-import org.mockito.Mockito
-import org.mockito.Mockito.verify
-import org.mockito.MockitoAnnotations
 import javax.inject.Inject
 
 
 @MediumTest
 @ExperimentalCoroutinesApi
 @RunWith(AndroidJUnit4::class)
-class CategoryListFragmentTest {
+class BrandListFragmentTest {
 
     @get:Rule
     var instantExecutorRule = InstantTaskExecutorRule()
 
     @Inject
-    lateinit var dataSource: IBrandCategoryDataSource<CategoryEntry>
-
-    @Inject
-    lateinit var navigator: Navigator
-
-    val sampleCategory1 = CategoryEntry(0, "Wagon")
-    val sampleCategory2 = CategoryEntry(1, "Locomotive")
-    val sampleCategoryList = mutableListOf(sampleCategory1, sampleCategory2)
+    lateinit var dataSource: IBrandCategoryDataSource<BrandEntry>
 
     @Before
     fun setUp() {
-        MockitoAnnotations.initMocks(this)
         val app = ApplicationProvider.getApplicationContext<AndroidTestApplication>()
         val component = app.appComponent as TestComponent
         component.inject(this)
@@ -67,8 +56,8 @@ class CategoryListFragmentTest {
     @Test
     fun withEmptyList_emptyScreenIsShown() {
         runBlockingTest {
-            (dataSource as FakeCategoryDataSource).setData(mutableListOf())
-            launchFragmentInContainer<CategoryListFragment>(Bundle(), R.style.AppTheme)
+            (dataSource as FakeBrandDataSource).setData(mutableListOf())
+            launchFragmentInContainer<BrandListFragment>(Bundle(), R.style.AppTheme)
 
             onView(withId(R.id.empty_text)).check(isVisible())
             onView(withId(R.id.empty_image)).check(isVisible())
@@ -81,8 +70,8 @@ class CategoryListFragmentTest {
     fun withSampleList_emptyScreenIsNotShown() {
         runBlockingTest {
             //Set some sample data
-            (dataSource as FakeCategoryDataSource).setData(sampleCategoryList)
-            launchFragmentInContainer<CategoryListFragment>(Bundle(), R.style.AppTheme)
+            (dataSource as FakeBrandDataSource).setData(sampleBrandList)
+            launchFragmentInContainer<BrandListFragment>(Bundle(), R.style.AppTheme)
 
             onView(withId(R.id.empty_text)).check(isGone())
             onView(withId(R.id.empty_image)).check(isGone())
@@ -90,26 +79,24 @@ class CategoryListFragmentTest {
         }
     }
 
-    //Click plus on the menu and verify that add child frag becomes visible with empty edittext
     @Test
     fun clickAdd_opensEmptyAddFragment() {
         runBlockingTest {
-            (dataSource as FakeCategoryDataSource).setData(sampleCategoryList)
-            val scenario = launchFragmentInContainer<CategoryListFragment>(Bundle(), R.style.AppTheme)
+            (dataSource as FakeBrandDataSource).setData(sampleBrandList)
+            val scenario = launchFragmentInContainer<BrandListFragment>(Bundle(), R.style.AppTheme)
 
             val context: Context = ApplicationProvider.getApplicationContext<AndroidTestApplication>()
             val addMenuItem = ActionMenuItem(context, 0, R.id.action_add, 0, 0, null)
-
             //Click on the add menu item
             scenario.onFragment { fragment ->
                 fragment.onOptionsItemSelected(addMenuItem)
             }
 
             //Check whether add category screen becomes visible
-            onView(withId(R.id.addCategory_editCatName)).check(matches(isDisplayed()))
-            onView(withId(R.id.addCategory_editCatName)).check(matches(withText("")))
-            onView(withId(R.id.addCategory_saveBtn)).check(matches(isDisplayed()))
-
+            onView(withId(R.id.addBrand_editBrandName)).check(matches(isDisplayed()))
+            onView(withId(R.id.addBrand_editBrandName)).check(matches(withText("")))
+            onView(withId(R.id.addBrand_image)).check(matches(isDisplayed()))
+            onView(withId(R.id.addBrand_saveBtn)).check(matches(isDisplayed()))
         }
     }
 
@@ -117,29 +104,17 @@ class CategoryListFragmentTest {
     @Test
     fun clickEditOnItem_opensAddFragmentFilled() {
         runBlockingTest {
-            (dataSource as FakeCategoryDataSource).setData(sampleCategoryList)
-            launchFragmentInContainer<CategoryListFragment>(Bundle(), R.style.AppTheme)
+            (dataSource as FakeBrandDataSource).setData(sampleBrandList)
+            launchFragmentInContainer<BrandListFragment>(Bundle(), R.style.AppTheme)
 
             onView(withId(R.id.list))
-                    .perform(actionOnItemAtPosition<RecyclerView.ViewHolder>(1, clickOnChildWithId(R.id.category_item_edit_icon)))
+                    .perform(actionOnItemAtPosition<RecyclerView.ViewHolder>(1, clickOnChildWithId(R.id.brand_item_edit_icon)))
 
-            onView(withId(R.id.addCategory_editCatName)).check(matches(isDisplayed()))
-            onView(withId(R.id.addCategory_editCatName)).check(matches(withText("Locomotive")))
-            onView(withId(R.id.addCategory_saveBtn)).check(matches(isDisplayed()))
-        }
-    }
-
-    //Click on train icon on a category and verify that navigator temps to launch TrainListFrag with correct inputs
-    @Test
-    fun clickTrainIconOnItem_launchesTrainListFragment() {
-        runBlockingTest {
-            (dataSource as FakeCategoryDataSource).setData(sampleCategoryList)
-            launchFragmentInContainer<CategoryListFragment>(Bundle(), R.style.AppTheme)
-
-            onView(withId(R.id.list))
-                    .perform(actionOnItemAtPosition<RecyclerView.ViewHolder>(1, clickOnChildWithId(R.id.category_item_train_icon)))
-
-            verify(navigator).launchTrainList_withThisCategory("Locomotive")
+            //Check whether add category screen becomes visible
+            onView(withId(R.id.addBrand_editBrandName)).check(matches(isDisplayed()))
+            onView(withId(R.id.addBrand_editBrandName)).check(matches(withText("MDN")))
+            onView(withId(R.id.addBrand_image)).check(matches(isDisplayed()))
+            onView(withId(R.id.addBrand_saveBtn)).check(matches(isDisplayed()))
         }
     }
 
@@ -147,9 +122,9 @@ class CategoryListFragmentTest {
     fun swipingItem_revealsDeleteConfirmation() {
         runBlockingTest {
             //Set some sample data
-            (dataSource as FakeCategoryDataSource).setData(sampleCategoryList)
+            (dataSource as FakeBrandDataSource).setData(sampleBrandList)
 
-            launchFragmentInContainer<CategoryListFragment>(Bundle(), R.style.AppTheme)
+            launchFragmentInContainer<BrandListFragment>(Bundle(), R.style.AppTheme)
 
             //Swipe an item
             onView(withId(R.id.list))
@@ -159,13 +134,6 @@ class CategoryListFragmentTest {
             onView(withId(R.id.confirm_delete_btn)).check(matches(withEffectiveVisibility(Visibility.VISIBLE)))
             onView(withId(R.id.cancel_btn)).check(matches(withEffectiveVisibility(Visibility.VISIBLE)))
             onView(withText(R.string.do_you_want_to_delete)).check(matches(withEffectiveVisibility(Visibility.VISIBLE)))
-        }
-    }
-
-    @After
-    fun validate() {
-        kotlin.runCatching {
-            Mockito.validateMockitoUsage()
         }
     }
 }
