+++ app/src/androidTest/java/com/canli/oya/traininventoryroom/fragments/TrainListFragmentTest.kt
@@ -1,13 +1,11 @@
-package com.canli.oya.traininventoryroom.ui.trains
+package com.canli.oya.traininventoryroom.fragments
 
 import android.os.Bundle
-import androidx.appcompat.view.menu.ActionMenuItem
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.fragment.app.testing.launchFragmentInContainer
 import androidx.recyclerview.widget.RecyclerView
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.espresso.Espresso.onView
-import androidx.test.espresso.action.ViewActions.click
 import androidx.test.espresso.action.ViewActions.swipeLeft
 import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.contrib.RecyclerViewActions
@@ -15,24 +13,17 @@ import androidx.test.espresso.matcher.ViewMatchers.*
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.filters.MediumTest
 import com.canli.oya.traininventoryroom.R
-import com.canli.oya.traininventoryroom.data.TrainEntry
-import com.canli.oya.traininventoryroom.data.source.FakeTrainDataSource
-import com.canli.oya.traininventoryroom.data.source.ITrainDataSource
+import com.canli.oya.traininventoryroom.data.source.*
 import com.canli.oya.traininventoryroom.di.AndroidTestApplication
 import com.canli.oya.traininventoryroom.di.TestComponent
-import com.canli.oya.traininventoryroom.ui.main.Navigator
+import com.canli.oya.traininventoryroom.ui.trains.TrainListFragment
 import com.canli.oya.traininventoryroom.utils.*
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runBlockingTest
-import org.junit.After
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
-import org.mockito.Mockito
-import org.mockito.Mockito.verify
-import org.mockito.MockitoAnnotations
-
 import javax.inject.Inject
 
 @MediumTest
@@ -46,17 +37,8 @@ class TrainListFragmentTest{
     @Inject
     lateinit var dataSource: ITrainDataSource
 
-    @Inject
-    lateinit var navigator: Navigator
-
-    val sampleTrain1 = TrainEntry(trainId = 0, trainName = "Red Wagon", categoryName = "Wagon", brandName = "Marklin")
-    val sampleTrain2 = TrainEntry(trainId = 1, trainName = "Blue Loco", categoryName = "Locomotif", brandName = "MDN")
-    val sampleTrain3 = TrainEntry(trainId = 2, trainName = "Gare", categoryName = "Accessoire", brandName = "Marklin")
-    val sampleTrainList = mutableListOf(sampleTrain1, sampleTrain2)
-
     @Before
     fun setUp() {
-        MockitoAnnotations.initMocks(this)
         val app = ApplicationProvider.getApplicationContext<AndroidTestApplication>()
         val component = app.appComponent as TestComponent
         component.inject(this)
@@ -191,54 +173,6 @@ class TrainListFragmentTest{
             onView(withId(R.id.confirm_delete_btn)).check(matches(withEffectiveVisibility(Visibility.VISIBLE)))
             onView(withId(R.id.cancel_btn)).check(matches(withEffectiveVisibility(Visibility.VISIBLE)))
             onView(withText(R.string.do_you_want_to_delete)).check(matches(withEffectiveVisibility(Visibility.VISIBLE)))
-        }
-    }
-
-    //Clicking a train item launches trainDetails fragment
-    @Test
-    fun clickingTrainItem_launchesTrainDetails() {
-        runBlockingTest {
-            //Set some sample data
-            (dataSource as FakeTrainDataSource).setData(sampleTrainList)
-
-            //Launch the fragment in ALL_TRAINS mode
-            val args = Bundle()
-            args.putString(INTENT_REQUEST_CODE, ALL_TRAIN)
-            launchFragmentInContainer<TrainListFragment>(args, R.style.AppTheme)
-
-            //Click on an item
-            onView(withId(R.id.list))
-                    .perform(RecyclerViewActions.actionOnItemAtPosition<RecyclerView.ViewHolder>(0, click()))
-
-            verify(navigator).launchTrainDetails(0)
-        }
-    }
-
-    //Clicking add menu on the item launches addTrainFragment
-    @Test
-    fun clickAddMenuItem_launchesAddTrainFrag() {
-        runBlockingTest {
-            //Set some sample data
-            (dataSource as FakeTrainDataSource).setData(mutableListOf())
-
-            //Launch the fragment in ALL_TRAINS mode
-            val args = Bundle()
-            args.putString(INTENT_REQUEST_CODE, ALL_TRAIN)
-            val scenario = launchFragmentInContainer<TrainListFragment>(args, R.style.AppTheme)
-            
-            val addMenuItem = ActionMenuItem(null, 0, R.id.action_add, 0, 0, null)
-            scenario.onFragment {
-                it.onOptionsItemSelected(addMenuItem)
-            }
-
-            verify(navigator).launchAddTrain()
-        }
-    }
-
-    @After
-    fun validate() {
-        kotlin.runCatching {
-            Mockito.validateMockitoUsage()
         }
     }
 }
